requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
requests==2.32.5
pillow==12.1.1
imageio-ffmpeg==0.6.0
